Updated Parsers (finished XYZ) & have successfully read data into Entity Framework classes & stored in the data. Also solved Foreign Key linking.

+++ DroneWebApp/DroneWebApp/Models/PointCloudXYZ.cs
@@ -21,7 +21,7 @@ namespace DroneWebApp.Models
         public Nullable<int> Red { get; set; }
         public Nullable<int> Green { get; set; }
         public Nullable<int> Blue { get; set; }
-        public Nullable<int> Intensity { get; set; }
+        public Nullable<double> Intensity { get; set; }
         public string FlightId { get; set; }
     
         public virtual DroneFlight DroneFlight { get; set; }
